pdf_search/requirements.txt
pymupdf>=1.24.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0
requests>=2.31.0
numpy>=1.24.0
langgraph>=0.0.20
langchain>=0.1.0
langchain-openai>=0.0.5
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
msgpack>=1.0.0
imageio-ffmpeg>=0.4.9
